# Performance optimizations for the inmem sinks key flattening